legal/AI/AI_jogi/requirements.txt
# Ai_jogi/requirements.txt (SERVER ONLY)

# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# LLM (Gemini)
google-generativeai==0.3.2

# Core
numpy==1.26.3
pandas==2.2.0
requests==2.31.0
python-dotenv==1.0.1
tqdm==4.66.1